Split custom hooks in different files
